+++ libcam/libcam_v4l2core/jpeg_decoder.c
@@ -1449,7 +1449,7 @@ int jpeg_init_decoder(int width, int height)
 #endif
 
 	/*alloc temp buffer*/
-	jpeg_ctx->tmp_frame = calloc(width*height*2, sizeof(uint8_t));
+    jpeg_ctx->tmp_frame = calloc((size_t)(width*height*2), sizeof(uint8_t));
 	if(jpeg_ctx->tmp_frame == NULL)
 	{
 		fprintf(stderr, "V4L2_CORE: FATAL memory allocation failure (jpeg_init_decoder): %s\n", strerror(errno));
@@ -1509,7 +1509,7 @@ int jpeg_decode(uint8_t *out_buf, uint8_t *in_buf, int size)
 	if(got_frame)
 	{
 #if LIBAVUTIL_VER_AT_LEAST(54,6)
-		av_image_copy_to_buffer(jpeg_ctx->tmp_frame, jpeg_ctx->pic_size,
+        av_image_copy_to_buffer(jpeg_ctx->tmp_frame, jpeg_ctx->pic_size,
                              (const uint8_t * const*) codec_data->picture->data, codec_data->picture->linesize,
                              codec_data->context->pix_fmt, jpeg_ctx->width, jpeg_ctx->height, 1);
 #else
